+++ lib/features/items/items.controller.dart
@@ -4,6 +4,7 @@ import 'package:liso/core/hive/models/item.hive.dart';
 
 import '../../core/utils/globals.dart';
 import '../drawer/drawer_widget.controller.dart';
+import '../pro/pro.controller.dart';
 import 'items.service.dart';
 
 class ItemsController extends GetxController with ConsoleMixin, StateMixin {
@@ -17,6 +18,11 @@ class ItemsController extends GetxController with ConsoleMixin, StateMixin {
   final sortOrder = LisoItemSortOrder.dateModifiedDescending.obs;
 
   // GETTERS
+  bool get itemLimitReached => data.length >= ProController.to.limits.items;
+
+  bool get protectedItemLimitReached =>
+      data.where((e) => e.protected).length >=
+      ProController.to.limits.protectedItems;
 
   // INIT
   @override