+++ lib/features/s3/explorer/s3_content_tile.controller.dart
@@ -4,10 +4,14 @@ import 'package:console_mixin/console_mixin.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:liso/core/hive/hive.service.dart';
+import 'package:liso/core/persistence/persistence.dart';
+import 'package:liso/features/main/main_screen.controller.dart';
 import 'package:liso/features/s3/explorer/s3_exporer_screen.controller.dart';
 import 'package:path/path.dart';
 import 'package:share_plus/share_plus.dart';
 
+import '../../../core/liso/liso.manager.dart';
 import '../../../core/liso/liso_paths.dart';
 import '../../../core/notifications/notifications.manager.dart';
 import '../../../core/services/cipher.service.dart';
@@ -15,18 +19,31 @@ import '../../../core/utils/file.util.dart';
 import '../../../core/utils/globals.dart';
 import '../../../core/utils/ui_utils.dart';
 import '../../../core/utils/utils.dart';
+import '../../app/routes.dart';
 import '../../attachments/attachments_screen.controller.dart';
 import '../model/s3_content.model.dart';
 import '../s3.service.dart';
 
 class S3ContentTileController extends GetxController
     with StateMixin, ConsoleMixin {
+  // VARIABLES
   String explorerType = '';
 
+  // PROPERTIES
+  final busy = false.obs;
+
+  // INIT
+  @override
+  void change(newState, {RxStatus? status}) {
+    busy.value = status?.isLoading ?? false;
+    super.change(newState, status: status);
+  }
+
+  // FUNCTIONS
   void share(S3Content content) async {
-    change(true, status: RxStatus.loading());
+    change('Sharing...', status: RxStatus.loading());
     final result = await S3Service.to.getPreSignedUrl(content.path);
-    change(false, status: RxStatus.success());
+    change('', status: RxStatus.success());
 
     if (result.isLeft) {
       return UIUtils.showSimpleDialog(
@@ -71,15 +88,17 @@ class S3ContentTileController extends GetxController
   }
 
   void askToImport(S3Content s3content) {
-    const content = Text(
-      'Are you sure you want to restore from this vault? \nYour current vault will be overwritten.',
+    if (busy.value) return;
+
+    final content = Text(
+      'Are you sure you want to restore from this backed-up vault ${s3content.name}?\n\nIf you choose to proceed: Your current vault will be overwritten alongside with all the items in it.',
     );
 
     Get.dialog(AlertDialog(
-      title: Text('restore'.tr),
+      title: Text("${'restore'.tr} Backup"),
       content: Utils.isDrawerExpandable
           ? content
-          : const SizedBox(
+          : SizedBox(
               width: 450,
               child: content,
             ),
@@ -96,20 +115,66 @@ class S3ContentTileController extends GetxController
     ));
   }
 
-  // void backup(S3Content content) async {
-  //   final result = await S3Service.to.backup(content);
+  void restore(S3Content content) async {
+    Get.back();
 
-  //   result.either(
-  //     (error) => UIUtils.showSimpleDialog(
-  //       'Error Backup',
-  //       error,
-  //     ),
-  //     (response) => console.info('backup success: $response'),
-  //   );
-  // }
+    change('Restoring...', status: RxStatus.loading());
+    // purge all items
+    await HiveService.to.purge();
+    // download chosen vault file
+    final downloadPath = join(LisoPaths.temp!.path, content.name);
 
-  void restore(S3Content content) {
-    // use S3Service.to.sync with a custom s3path
+    final downloadResult = await S3Service.to.downloadFile(
+      s3Path: content.path,
+      filePath: downloadPath,
+    );
+
+    if (downloadResult.isLeft) {
+      change('Failed to download', status: RxStatus.success());
+
+      return UIUtils.showSimpleDialog(
+        'Failed To Download',
+        '${downloadResult.left} -> download()',
+      );
+    }
+
+    final vaultFile = downloadResult.right;
+
+    // re-upload to overwrite current vault file
+    final uploadResult = await S3Service.to.uploadFile(
+      vaultFile,
+      s3Path: S3Service.to.vaultPath,
+    );
+
+    if (uploadResult.isLeft) {
+      change('Failed to upload', status: RxStatus.success());
+
+      return UIUtils.showSimpleDialog(
+        'Upload Failed',
+        'Error: ${uploadResult.left}',
+      );
+    }
+
+    // parse vault file
+    final vault = await LisoManager.parseVaultFile(
+      vaultFile,
+      cipherKey: Persistence.to.cipherKey,
+    );
+
+    // import vault object & reload
+    await LisoManager.importVault(
+      vault,
+      cipherKey: Persistence.to.cipherKey,
+    );
+
+    MainScreenController.to.load();
+
+    NotificationsManager.notify(
+      title: 'Vault Restored',
+      body: '${content.name} successfully restored!',
+    );
+
+    Get.offNamedUntil(Routes.main, (route) => false);
   }
 
   // TODO: confirmation dialog
@@ -121,7 +186,7 @@ class S3ContentTileController extends GetxController
 
     void _delete() async {
       Get.back();
-      change(true, status: RxStatus.loading());
+      change('Deleting...', status: RxStatus.loading());
       final result = await S3Service.to.remove(content);
 
       if (result.isLeft) {
@@ -138,7 +203,7 @@ class S3ContentTileController extends GetxController
         body: content.name,
       );
 
-      change(false, status: RxStatus.success());
+      change('false', status: RxStatus.success());
       await S3ExplorerScreenController.to.reload();
     }
 
@@ -192,7 +257,7 @@ class S3ContentTileController extends GetxController
   }
 
   void _download(S3Content content) async {
-    change(true, status: RxStatus.loading());
+    change('Downloading...', status: RxStatus.loading());
     final downloadPath = join(LisoPaths.temp!.path, content.name);
 
     final result = await S3Service.to.downloadFile(
@@ -201,7 +266,7 @@ class S3ContentTileController extends GetxController
     );
 
     if (result.isLeft) {
-      change(false, status: RxStatus.success());
+      change('', status: RxStatus.success());
 
       return UIUtils.showSimpleDialog(
         'Failed To Download',
@@ -227,7 +292,7 @@ class S3ContentTileController extends GetxController
       );
 
       Globals.timeLockEnabled = true; // re-enable
-      return change(false, status: RxStatus.success());
+      return change('', status: RxStatus.success());
     }
 
     // choose directory and export file
@@ -250,6 +315,6 @@ class S3ContentTileController extends GetxController
       body: fileName,
     );
 
-    change(false, status: RxStatus.success());
+    change('', status: RxStatus.success());
   }
 }